+++ src/reactive/index.ts
@@ -27,8 +27,9 @@ export { immutables, isNonReactive, registerNativeReactivity } from './non-react
 export { getActiveProjection, project } from './project'
 export { isReactive, ReactiveBase, reactive, unwrap } from './proxy'
 export { organize, organized } from './record'
-export { scan, type ScanResult } from './buffer'
+export { scan, type ScanResult, resolve } from './buffer'
 export { Register, register } from './register'
+export { getActiveEffect, effectAggregator } from './effect-context'
 export {
 	type DependencyAccess,
 	type Evolution,